+++ src/modules/layerInfo/i18n/layerInfo.provider.js
@@ -4,14 +4,14 @@ export const provide = (lang) => {
 		case 'en':
 			return {
 				//the first part of the snake_case key should be the name of the related module
-				layerinfo_empty_layerInfo: 'No information for the layer found',
-				layerinfo_layerInfo_response_error: 'The layer-information could not be loaded'
+				layerinfo_empty_layerInfo: 'No Layer Information available',
+				layerinfo_layerInfo_response_error: 'The Layer Information could not be loaded'
 			};
 
 		case 'de':
 			return {
 				//the first part of the snake_case key should be the name of the related module
-				layerinfo_empty_layerInfo: 'Keine Informationen für die Ebene gefunden',
+				layerinfo_empty_layerInfo: 'Kein Ebenen-Information verfügbar',
 				layerinfo_layerInfo_response_error: 'Die Ebenen-Information konnte nicht geladen werden'
 			};
